baselines/g2_reader/requirements.txt
acres==0.5.0
aiodns
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asttokens==3.0.1
async-timeout
attrs
azure-core==1.38.0
beautifulsoup4==4.14.3
brotli==1.2.0
brotlicffi
bs4==0.0.2
cachetools==6.2.1
certifi
cffi
chardet==5.2.0
charset-normalizer
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
distro
etelemetry==0.3.1
exceptiongroup
executing==2.2.1
fast-langdetect==1.0.0
fastapi==0.118.3
fasttext-predict==0.9.2.4
filelock
fonttools==4.61.1
frozenlist
fsspec
gmpy2
greenlet
h11
hf-xet==1.1.10
html2text==2025.4.15
httpcore
httplib2==0.31.0
httpx
huggingface-hub==0.35.3
idna==3.10
importlib_metadata
importlib_resources==6.5.2
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2
jiter
joblib==1.5.2
jsonlines==4.0.0
jsonpatch
jsonpickle==4.1.1
jsonpointer
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain
langchain-core
langchain-text-splitters
langsmith
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.2
magic-pdf==1.3.12
MarkupSafe
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath
multidict
narwhals==2.14.0
networkx
nibabel==5.3.2
nipype==1.10.0
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
openai
opencv-python-headless==4.11.0.86
opentelemetry-api
orjson
packaging==23.2
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
pdf2image
pexpect==4.9.0
pillow
plotly==6.5.0
propcache
prov==2.1.1
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pycares
pycparser
pydantic
pydantic_core
pydot==4.0.1
PyMuPDF==1.26.6
pyparsing==3.2.5
PyPDF2
PySocks
pytesseract
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
pyxnat==1.6.4
PyYAML
rdflib==7.4.0
referencing==0.37.0
regex
requests
requests-toolbelt
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
simplejson==3.20.2
six==1.17.0
sniffio
soupsieve==2.8.1
SQLAlchemy
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken
tokenizers==0.22.1
torch
torchaudio==2.5.1
torchvision==0.20.1
tqdm
traitlets==5.14.3
traits==7.0.2
transformers==4.57.0
triton==3.4.0
typing-inspection
typing_extensions
tzdata==2025.2
urllib3
uvicorn==0.37.0
uvloop
vllm==0.8.5
yarl
zipp
zstandard
